requirements.txt
pymobiledevice3>=9.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
pywebview>=5.0
